+++ apps/login/src/lib/server/password.ts
@@ -14,7 +14,8 @@ import {
   setPassword,
   setUserPassword,
 } from "@/lib/zitadel";
-import { create, createServerTransport } from "@zitadel/client";
+import { create } from "@zitadel/client";
+import { createServerTransport } from "@zitadel/client/node";
 import { createUserServiceClient } from "@zitadel/client/v2";
 import {
   Checks,
