+++ packages/zitadel-client/package.json
@@ -27,6 +27,11 @@
       "types": "./dist/v3alpha.d.ts",
       "import": "./dist/v3alpha.js",
       "require": "./dist/v3alpha.cjs"
+    },
+    "./node": {
+      "types": "./dist/node.d.ts",
+      "import": "./dist/node.js",
+      "require": "./dist/node.cjs"
     }
   },
   "files": [
