+++ packages/zitadel-client/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig, Options } from "tsup";
 
 export default defineConfig((options: Options) => ({
-  entry: ["src/index.ts", "src/v1.ts", "src/v2.ts", "src/v3alpha.ts"],
+  entry: ["src/index.ts", "src/v1.ts", "src/v2.ts", "src/v3alpha.ts", "src/node.ts"],
   format: ["esm", "cjs"],
   treeshake: false,
   splitting: true,
